fix size_t
git-svn-id: http://ltt.polymtl.ca/svn@1827 04897980-b3bd-0310-b5e0-8ef037075253

+++ ltt-usertrace/ltt/ltt-usertrace-fast.h
@@ -349,7 +349,7 @@ static inline void * __attribute__((no_instrument_function)) ltt_reserve_slot(
 															struct ltt_trace_info *trace,
 															struct ltt_buf *ltt_buf,
 															unsigned int data_size,
-															unsigned int *slot_size,
+															size_t *slot_size,
 															uint64_t *tsc,
 															size_t *before_hdr_pad,
 															size_t *after_hdr_pad,
